requirements.txt
openai>=1.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0


pillow~=11.3.0
pyowm~=3.3.0